pyproject.toml
[project]
name = "py-uds-ecu-address"
description = "Collection of UDS addresses for various ECUs."
readme = "README.rst"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking",
]
keywords = [
    "uds",
    "unified-diagnostic-services",
    "automotive",
    "on-board-diagnostic",
    "obd",
    "vehicle-diagnostic",
    "iso14229",
    "iso-14229",
    "iso15765",
    "iso-15765",
]
authors = [
    {name="Maciej Dąbrowski", email="maciej.dabrowski@software-quality.pl"},
    {name="UDS Package Development", email="uds-package-development@googlegroups.com"},
]
maintainers = [
    {name="Maciej Dąbrowski", email="maciej.dabrowski@software-quality.pl"},
    {name="UDS Package Development", email="uds-package-development@googlegroups.com"},
]
dependencies = [
    "py-uds >= 1.1.0"
]
dynamic = ["version"]


[project.urls]
"Homepage" = "https://github.com/mdabrowski1990/uds"
"User Documentation" = "https://uds.readthedocs.io"
"Repository" = "https://github.com/mdabrowski1990/uds-ecu-address"
"Issues Tracking System" = "https://github.com/mdabrowski1990/uds-ecu-address/issues"
"Releases and Changelogs" = "https://github.com/mdabrowski1990/uds-ecu-address/releases"
"PyPI" = "https://pypi.org/project/py-uds-ecu-address/"
"Code Download" = "https://github.com/mdabrowski1990/uds-ecu-address/archive/refs/heads/main.zip"


[build-system]
requires = [
    "setuptools >= 42",
    "wheel"
]
build-backend = "setuptools.build_meta"


[tool.setuptools.dynamic]
version = {attr = "uds.__version__"}


[tool.setuptools.package-data]
"*" = [
    "*.md",
    "*.rst",
    "*.txt",
    "*.toml"
]


[tool.setuptools.packages.find]
include = [
    "uds_ecu_address.*"
]


[tool.isort]
src_paths = ["uds_ecu_address"]
line_length = 120
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
multi_line_output = 3
